requirements.txt
openai
python-dotenv
requests
pydantic
icalendar
fastapi
uvicorn
mcp
streamlit
folium
streamlit-folium
